pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-radar"
version = "0.1.0"
description = "Turn Reddit into your daily, goal-specific research radar"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Reddit Radar Contributors"}
]
keywords = ["reddit", "research", "ai", "llm", "insights"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "praw>=7.7.0",
    "openai>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
reddit-radar = "main:cli"

[project.urls]
Homepage = "https://github.com/skaczmarczyk/reddit-radar"
Documentation = "https://github.com/skaczmarczyk/reddit-radar#readme"
Repository = "https://github.com/skaczmarczyk/reddit-radar"
Issues = "https://github.com/skaczmarczyk/reddit-radar/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
